sync repo

+++ custom_components/hacs/handler/download.py
@@ -1,23 +1,22 @@
 """Download."""
+import os
 import gzip
-import logging
 import shutil
 
 import aiofiles
 import async_timeout
-
+from integrationhelper import Logger
 import backoff
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from ..exceptions import HacsNotSoBasicException
-
-_LOGGER = logging.getLogger("custom_components.hacs.download")
+from ..hacsbase.exceptions import HacsNotSoBasicException
 
 
-@backoff.on_exception(backoff.expo, Exception, max_tries=3)
+@backoff.on_exception(backoff.expo, Exception, max_tries=5)
 async def async_download_file(hass, url):
     """
     Download files, and return the content.
     """
+    logger = Logger("hacs.download.downloader")
     if url is None:
         return
 
@@ -25,7 +24,7 @@ async def async_download_file(hass, url):
     if "tags/" in url:
         url = url.replace("tags/", "")
 
-    _LOGGER.debug("Donwloading %s", url)
+    logger.debug(f"Donwloading {url}")
 
     result = None
 
@@ -34,10 +33,7 @@ async def async_download_file(hass, url):
 
         # Make sure that we got a valid result
         if request.status == 200:
-            if url.endswith(".gz"):
-                result = await request.read()
-            else:
-                result = await request.text()
+            result = await request.read()
         else:
             raise HacsNotSoBasicException(
                 "Got status code {} when trying to download {}".format(
@@ -50,14 +46,8 @@ async def async_download_file(hass, url):
 
 async def async_save_file(location, content):
     """Save files."""
-    if "-bundle" in location:
-        location = location.replace("-bundle", "")
-    if "lovelace-" in location.split("/")[-1]:
-        search = location.split("/")[-1]
-        replace = search.replace("lovelace-", "")
-        location = location.replace(search, replace)
-
-    _LOGGER.debug("Saving %s", location)
+    logger = Logger("hacs.download.save")
+    logger.debug(f"Saving {location}")
     mode = "w"
     encoding = "utf-8"
     errors = "ignore"
@@ -74,12 +64,15 @@ async def async_save_file(location, content):
             await outfile.write(content)
             outfile.close()
 
+        # Create gz for .js files
+        if os.path.isfile(location):
+            if location.endswith(".js") or location.endswith(".css"):
+                with open(location, "rb") as f_in:
+                    with gzip.open(location + ".gz", "wb") as f_out:
+                        shutil.copyfileobj(f_in, f_out)
+
     except Exception as error:  # pylint: disable=broad-except
         msg = "Could not write data to {} - {}".format(location, error)
-        _LOGGER.debug(msg)
+        logger.debug(msg)
 
-    # Create gz for .js files
-    if location.endswith(".js") or location.endswith(".css"):
-        with open(location, "rb") as f_in:
-            with gzip.open(location + ".gz", "wb") as f_out:
-                shutil.copyfileobj(f_in, f_out)
+    return os.path.exists(location)
